requirements.txt
spotipy
streamlit
pandas
matplotlib
numpy
python-dotenv
uvicorn